backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
psycopg[binary]>=3.2.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
